opaque pointers : refactor *_type functions
Checking the structure name for "opencl.image" is similar to the
utility functions is_image_type() and is_sampler_type().  So make
this block of code into is_ocl_image_type().

In clang 15 the use of PointerElementType() is flagged as
deprecated.  So change the is_image_type() and is_sampler_type()
to how is_ocl_image_type() is handled, by passing in the
ValueType pointer.

To caller of is_image_type() and is_sampler_type(),
pocl_llvm_get_kernels_metadata(), rework the loop over the args
to  be similar to how fixOpenCLimageArguements() is structured.

Add ifdef for LLVM_OPAQUE_POINTERS to pick between using
getPointerElementType() or getParamByValType()

Signed-off-by: Tom Rix <trix@redhat.com>